Fix for time_counter

+++ sw/time/include/time_counter.h
@@ -19,7 +19,7 @@ typedef struct
 	uint8_t hour;
 	uint8_t minute;
 	uint8_t second;
-	uint8_t msecond;
+	uint16_t msecond;
 }TimeItem;
